Price on groceries from decimal to int, deleting commentresource

+++ SwiftCart/database/migrations/2025_03_24_201638_create_products_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('name');
             $table->string('brand');
             $table->string('description');
-            $table->decimal('price');
+            $table->unsignedBigInteger('price');
             $table->unsignedBigInteger('map_segment_id');
             $table->timestamps();
